requirements.txt
Flask
twilio
google-generativeai
gunicorn